feat: Add user data to submissions and choiceguide results
