+++ apps/api-server/src/models/ChoicesGuideResult.js
@@ -76,7 +76,18 @@ module.exports = function( db, sequelize, DataTypes ) {
         include: [
           {
             model: db.User,
-            attributes: ['role', 'displayName', 'nickName', 'name', 'email'],
+            attributes: [
+              'role',
+              'displayName',
+              'nickName',
+              'name',
+              'displayName',
+              'email',
+              'phonenumber',
+              'address',
+              'city',
+              'postcode'
+            ],
           },
         ],
       },
