+++ apps/api-server/src/models/Submission.js
@@ -40,7 +40,18 @@ module.exports = function (db, sequelize, DataTypes) {
         include: [
           {
             model: db.User,
-            attributes: ['role', 'displayName', 'nickName', 'name', 'email'],
+            attributes: [
+              'role',
+              'displayName',
+              'nickName',
+              'name',
+              'displayName',
+              'email',
+              'phonenumber',
+              'address',
+              'city',
+              'postcode'
+            ],
           },
         ],
       },
@@ -64,6 +75,7 @@ module.exports = function (db, sequelize, DataTypes) {
 
   Submission.associate = function (models) {
     this.belongsTo(models.Widget);
+    Submission.belongsTo(models.User, { onDelete: 'CASCADE' });
   };
 
   return Submission;
